+++ internal/social/services/router.go
@@ -40,7 +40,7 @@ func (r *serviceRouter) NotifyServer(_ context.Context, res *service.NotifyServe
 
 func (r *serviceRouter) Join(_ context.Context, res *service.JoinRequest) (*emptypb.Empty, error) {
 	for _, s := range r.services {
-		u, err := s.GetUserID(int64(res.UserId))
+		u, err := s.GetUserID(res.UserId)
 		if err != nil {
 			return nil, UserNotFound
 		}
